Create utils/configurations module
